llvm-pass: added custom graph traits
In LLVM, only `DOTGraphTraits<const Function *>` is provided. However,
casting `Function` to `const Function` when calling `WriteGraph` can
cause a segfault (this is not a safe cast). Unfortunately, dropping the
cast means that the `DefaultDOTGraphTraits` class is used, which always
produces the empty string for a node label, which breaks distance
calculations.

Solution is to add a new `DOTGraphTraits<Function *>` so that undefined
casts are avoided.