Fix raw_fd_ostream constructor call
LLVM 4.x does not have this specific constructor call yet.
However, the constructor with additional sys::fs::OpenFlags
is available in 4.x, and still exists in trunk.

Furthermore, the constant sys::fs::F_None still exists for compatibility
in trunk as well, so these changes should not affect builds with more
recent LLVM versions.

Use this call instead with the original 2-argument constructor
default value to allow aflgo to build with LLVM 4.0 out of the box.