Fix node naming, and help
Only looking for "{name" could result in "{bla:23" matching to
"{bla:230", which resulted in wrong results for cfgs. This was
previously not spotted, because the old distance files were used while
testing, which was a result of a silent failure fixed in (4d66b59692b3).

This commit also fixes an issue with displaying the help text.